requirements.txt
streamlit>=1.24.0
langchain>=0.0.207
langchain_core>=0.0.5
langchain_openai>=0.0.2
langchain_community>=0.0.5
openai>=0.28.0
python-dotenv>=1.0.0
requests>=2.28.2
regex>=2022.10.31
google-generativeai>=0.3.0  # For Google AI Studio
googlesearch-python>=1.2.3  # For web search
beautifulsoup4>=4.12.0  # For parsing search results
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5